@@ -1,13 +1,20 @@
-module github.com/vonage/vonage-go-sdk
+module github.com/adevinta/vonage-go-sdk
 
-go 1.13
+go 1.23
 
 require (
 	github.com/antihax/optional v1.0.0
-	github.com/golang-jwt/jwt v3.2.1+incompatible // indirect
+	github.com/golang-jwt/jwt v3.2.1+incompatible
 	github.com/google/uuid v1.1.1
 	github.com/jarcoal/httpmock v1.0.4
 	github.com/spf13/cobra v0.0.6
-	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
+)
+
+require (
+	github.com/golang/protobuf v1.3.1 // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/net v0.33.0 // indirect
+	google.golang.org/appengine v1.4.0 // indirect
 )
