@@ -11,17 +11,18 @@
 package numberinsight
 
 import (
-	_context "context"
-	"github.com/antihax/optional"
-	_ioutil "io/ioutil"
-	_nethttp "net/http"
-	_neturl "net/url"
+	"context"
+	"io"
+	"net/http"
+	"net/url"
 	"strings"
+
+	"github.com/antihax/optional"
 )
 
 // Linger please
 var (
-	_ _context.Context
+	_ context.Context
 )
 
 // DefaultApiService DefaultApi service
@@ -37,18 +38,19 @@ type GetNumberInsightAdvancedOpts struct {
 /*
 GetNumberInsightAdvanced Advanced Number Insight (sync)
 Provides [advanced number insight](/number-insight/overview#basic-standard-and-advanced-apis) information about a number synchronously, in the same way that the basic and standard endpoints do.  Nexmo recommends accessing the Advanced API **asynchronously** using the &#x60;/advanced/async&#x60; endpoint, to avoid timeouts.  Note that this endpoint also supports &#x60;POST&#x60; requests.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The format of the response
- * @param number A single phone number that you need insight about in national or international format.
- * @param optional nil or *GetNumberInsightAdvancedOpts - Optional Parameters:
- * @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
- * @param "Cnam" (optional.Bool) -  Indicates if the name of the person who owns the phone number should be looked up and returned in the response. Set to true to receive phone number owner name in the response. This features is available for US numbers only and incurs an additional charge.
- * @param "Ip" (optional.String) -  This parameter is deprecated as we are no longer able to retrieve reliable IP data globally from carriers.
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The format of the response
+  - @param number A single phone number that you need insight about in national or international format.
+  - @param optional nil or *GetNumberInsightAdvancedOpts - Optional Parameters:
+  - @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
+  - @param "Cnam" (optional.Bool) -  Indicates if the name of the person who owns the phone number should be looked up and returned in the response. Set to true to receive phone number owner name in the response. This features is available for US numbers only and incurs an additional charge.
+  - @param "Ip" (optional.String) -  This parameter is deprecated as we are no longer able to retrieve reliable IP data globally from carriers.
+
 @return NiResponseJsonAdvanced
 */
-func (a *DefaultApiService) GetNumberInsightAdvanced(ctx _context.Context, format string, number string, localVarOptionals *GetNumberInsightAdvancedOpts) (NiResponseJsonAdvanced, *_nethttp.Response, error) {
+func (a *DefaultApiService) GetNumberInsightAdvanced(ctx context.Context, format string, number string, localVarOptionals *GetNumberInsightAdvancedOpts) (NiResponseJsonAdvanced, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = http.MethodGet
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -58,11 +60,11 @@ func (a *DefaultApiService) GetNumberInsightAdvanced(ctx _context.Context, forma
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/advanced/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	localVarQueryParams.Add("number", parameterToString(number, ""))
 	if localVarOptionals != nil && localVarOptionals.Country.IsSet() {
@@ -125,7 +127,7 @@ func (a *DefaultApiService) GetNumberInsightAdvanced(ctx _context.Context, forma
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -161,19 +163,20 @@ type GetNumberInsightAsyncOpts struct {
 /*
 GetNumberInsightAsync Advanced Number Insight (async)
 Provides [advanced number insight](/number-insight/overview#basic-standard-and-advanced-apis) number information **asynchronously** using the URL specified in the &#x60;callback&#x60; parameter. Nexmo recommends asynchronous use of the Number Insight Advanced API, to avoid timeouts.  Note that this endpoint also supports &#x60;POST&#x60; requests.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The format of the response
- * @param callback The callback URL
- * @param number A single phone number that you need insight about in national or international format.
- * @param optional nil or *GetNumberInsightAsyncOpts - Optional Parameters:
- * @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
- * @param "Cnam" (optional.Bool) -  Indicates if the name of the person who owns the phone number should be looked up and returned in the response. Set to true to receive phone number owner name in the response. This features is available for US numbers only and incurs an additional charge.
- * @param "Ip" (optional.String) -  This parameter is deprecated as we are no longer able to retrieve reliable IP data globally from carriers.
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The format of the response
+  - @param callback The callback URL
+  - @param number A single phone number that you need insight about in national or international format.
+  - @param optional nil or *GetNumberInsightAsyncOpts - Optional Parameters:
+  - @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
+  - @param "Cnam" (optional.Bool) -  Indicates if the name of the person who owns the phone number should be looked up and returned in the response. Set to true to receive phone number owner name in the response. This features is available for US numbers only and incurs an additional charge.
+  - @param "Ip" (optional.String) -  This parameter is deprecated as we are no longer able to retrieve reliable IP data globally from carriers.
+
 @return NiResponseAsync
 */
-func (a *DefaultApiService) GetNumberInsightAsync(ctx _context.Context, format string, callback string, number string, localVarOptionals *GetNumberInsightAsyncOpts) (NiResponseAsync, *_nethttp.Response, error) {
+func (a *DefaultApiService) GetNumberInsightAsync(ctx context.Context, format string, callback string, number string, localVarOptionals *GetNumberInsightAsyncOpts) (NiResponseAsync, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = http.MethodGet
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -183,11 +186,11 @@ func (a *DefaultApiService) GetNumberInsightAsync(ctx _context.Context, format s
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/advanced/async/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	localVarQueryParams.Add("callback", parameterToString(callback, ""))
 	localVarQueryParams.Add("number", parameterToString(number, ""))
@@ -251,7 +254,7 @@ func (a *DefaultApiService) GetNumberInsightAsync(ctx _context.Context, format s
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -285,16 +288,17 @@ type GetNumberInsightBasicOpts struct {
 /*
 GetNumberInsightBasic Basic Number Insight
 Provides [basic number insight](/number-insight/overview#basic-standard-and-advanced-apis) information about a number.  Note that this endpoint also supports &#x60;POST&#x60; requests.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The format of the response
- * @param number A single phone number that you need insight about in national or international format.
- * @param optional nil or *GetNumberInsightBasicOpts - Optional Parameters:
- * @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The format of the response
+  - @param number A single phone number that you need insight about in national or international format.
+  - @param optional nil or *GetNumberInsightBasicOpts - Optional Parameters:
+  - @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
+
 @return NiResponseJsonBasic
 */
-func (a *DefaultApiService) GetNumberInsightBasic(ctx _context.Context, format string, number string, localVarOptionals *GetNumberInsightBasicOpts) (NiResponseJsonBasic, *_nethttp.Response, error) {
+func (a *DefaultApiService) GetNumberInsightBasic(ctx context.Context, format string, number string, localVarOptionals *GetNumberInsightBasicOpts) (NiResponseJsonBasic, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = http.MethodGet
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -304,11 +308,11 @@ func (a *DefaultApiService) GetNumberInsightBasic(ctx _context.Context, format s
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/basic/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	localVarQueryParams.Add("number", parameterToString(number, ""))
 	if localVarOptionals != nil && localVarOptionals.Country.IsSet() {
@@ -365,7 +369,7 @@ func (a *DefaultApiService) GetNumberInsightBasic(ctx _context.Context, format s
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -400,17 +404,18 @@ type GetNumberInsightStandardOpts struct {
 /*
 GetNumberInsightStandard Standard Number Insight
 Provides [standard number insight](/number-insight/overview#basic-standard-and-advanced-apis) information about a number.  Note that this endpoint also supports &#x60;POST&#x60; requests.
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The format of the response
- * @param number A single phone number that you need insight about in national or international format.
- * @param optional nil or *GetNumberInsightStandardOpts - Optional Parameters:
- * @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
- * @param "Cnam" (optional.Bool) -  Indicates if the name of the person who owns the phone number should be looked up and returned in the response. Set to true to receive phone number owner name in the response. This features is available for US numbers only and incurs an additional charge.
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The format of the response
+  - @param number A single phone number that you need insight about in national or international format.
+  - @param optional nil or *GetNumberInsightStandardOpts - Optional Parameters:
+  - @param "Country" (optional.String) -  If a number does not have a country code or is uncertain, set the two-character country code. This code must be in ISO 3166-1 alpha-2 format and in upper case. For example, GB or US. If you set country and number is already in [E.164](https://en.wikipedia.org/wiki/E.164) format, country must match the country code in number.
+  - @param "Cnam" (optional.Bool) -  Indicates if the name of the person who owns the phone number should be looked up and returned in the response. Set to true to receive phone number owner name in the response. This features is available for US numbers only and incurs an additional charge.
+
 @return NiResponseJsonStandard
 */
-func (a *DefaultApiService) GetNumberInsightStandard(ctx _context.Context, format string, number string, localVarOptionals *GetNumberInsightStandardOpts) (NiResponseJsonStandard, *_nethttp.Response, error) {
+func (a *DefaultApiService) GetNumberInsightStandard(ctx context.Context, format string, number string, localVarOptionals *GetNumberInsightStandardOpts) (NiResponseJsonStandard, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = http.MethodGet
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -420,11 +425,11 @@ func (a *DefaultApiService) GetNumberInsightStandard(ctx _context.Context, forma
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/standard/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	localVarQueryParams.Add("number", parameterToString(number, ""))
 	if localVarOptionals != nil && localVarOptionals.Country.IsSet() {
@@ -484,7 +489,7 @@ func (a *DefaultApiService) GetNumberInsightStandard(ctx _context.Context, forma
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
