@@ -12,11 +12,10 @@ package verify
 
 import (
 	"bytes"
-	_context "context"
-	"io/ioutil"
-	_ioutil "io/ioutil"
-	_nethttp "net/http"
-	_neturl "net/url"
+	"context"
+	"io"
+	"net/http"
+	"net/url"
 	"reflect"
 	"strings"
 
@@ -25,7 +24,7 @@ import (
 
 // Linger please
 var (
-	_ _context.Context
+	_ context.Context
 )
 
 // DefaultApiService DefaultApi service
@@ -39,19 +38,20 @@ type VerifyCheckOpts struct {
 /*
 VerifyCheck Verify Check
 Use Verify check to confirm that the PIN you received from your user matches the one sent by Nexmo in your Verify request.  1. Send the verification &#x60;code&#x60; that your user supplied, with the corresponding &#x60;request_id&#x60; from the Verify request. 2. Check the &#x60;status&#x60; of the response to determine if the code the user supplied matches the one sent by Nexmo.  *Note that this endpoint is available by &#x60;GET&#x60; request as well as &#x60;POST&#x60;.*
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The response format.
- * @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
- * @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
- * @param requestId The Verify request to check. This is the `request_id` you received in the response to the Verify request.
- * @param code The verification code entered by your user.
- * @param optional nil or *VerifyCheckOpts - Optional Parameters:
- * @param "IpAddress" (optional.String) -  (This field is no longer used)
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The response format.
+  - @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
+  - @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
+  - @param requestId The Verify request to check. This is the `request_id` you received in the response to the Verify request.
+  - @param code The verification code entered by your user.
+  - @param optional nil or *VerifyCheckOpts - Optional Parameters:
+  - @param "IpAddress" (optional.String) -  (This field is no longer used)
+
 @return CheckResponse
 */
-func (a *DefaultApiService) VerifyCheck(ctx _context.Context, format string, apiKey string, apiSecret string, requestId string, code string, localVarOptionals *VerifyCheckOpts) (CheckResponse, *_nethttp.Response, error) {
+func (a *DefaultApiService) VerifyCheck(ctx context.Context, format string, apiKey string, apiSecret string, requestId string, code string, localVarOptionals *VerifyCheckOpts) (CheckResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -61,11 +61,11 @@ func (a *DefaultApiService) VerifyCheck(ctx _context.Context, format string, api
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/check/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 	if strlen(requestId) > 32 {
 		return localVarReturnValue, nil, reportError("requestId must have less than 32 elements")
 	}
@@ -110,11 +110,11 @@ func (a *DefaultApiService) VerifyCheck(ctx _context.Context, format string, api
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 
 	// hack to reinstate the body in case we need it
-	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -143,17 +143,18 @@ func (a *DefaultApiService) VerifyCheck(ctx _context.Context, format string, api
 /*
 VerifyControl Verify Control
 Control the progress of your Verify requests. To cancel an existing Verify request, or to trigger the next verification event:   1. Send a Verify control request with the appropriate command (&#x60;cmd&#x60;) for what you want to achieve.  2. Check the &#x60;status&#x60; in the response.   *Note that this endpoint is available by &#x60;GET&#x60; request as well as &#x60;POST&#x60;.*
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The response format.
- * @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
- * @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
- * @param requestId The `request_id` you received in the response to the Verify request.
- * @param cmd The possible commands are `cancel` to request cancellation of the verification process, or `trigger_next_event` to advance  to the next verification event (if any). Cancellation is only possible 30 seconds after the start of the verification request and before the second event (either TTS or SMS) has taken place.
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The response format.
+  - @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
+  - @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
+  - @param requestId The `request_id` you received in the response to the Verify request.
+  - @param cmd The possible commands are `cancel` to request cancellation of the verification process, or `trigger_next_event` to advance  to the next verification event (if any). Cancellation is only possible 30 seconds after the start of the verification request and before the second event (either TTS or SMS) has taken place.
+
 @return ControlResponse
 */
-func (a *DefaultApiService) VerifyControl(ctx _context.Context, format string, apiKey string, apiSecret string, requestId string, cmd string) (ControlResponse, *_nethttp.Response, error) {
+func (a *DefaultApiService) VerifyControl(ctx context.Context, format string, apiKey string, apiSecret string, requestId string, cmd string) (ControlResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -163,11 +164,11 @@ func (a *DefaultApiService) VerifyControl(ctx _context.Context, format string, a
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/control/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{"application/x-www-form-urlencoded"}
@@ -200,11 +201,11 @@ func (a *DefaultApiService) VerifyControl(ctx _context.Context, format string, a
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 
 	// hack to reinstate the body in case we need it
-	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -244,25 +245,26 @@ type VerifyRequestOpts struct {
 /*
 VerifyRequest Request a Verification
 Use Verify request to generate and send a PIN to your user:  1. Create a request to send a verification code to your user.  2. Check the &#x60;status&#x60; field in the response to ensure that your request was successful (zero is success).  3. Use the &#x60;request_id&#x60; field in the response for the Verify check.  *Note that this endpoint is available by &#x60;GET&#x60; request as well as &#x60;POST&#x60;.*
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The response format.
- * @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
- * @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
- * @param number The mobile or landline phone number to verify. Unless you are setting `country` explicitly, this number must be in [E.164](https://en.wikipedia.org/wiki/E.164) format.
- * @param brand An 18-character alphanumeric string you can use to personalize the verification request SMS body, to help users identify your company or application name. For example: \\\"Your `Acme Inc` PIN is ...\\\"
- * @param optional nil or *VerifyRequestOpts - Optional Parameters:
- * @param "Country" (optional.String) -  If you do not provide `number` in international format or you are not sure if `number` is correctly formatted, specify the two-character country code in `country`. Verify will then format the number for you.
- * @param "SenderId" (optional.String) -  An 11-character alphanumeric string that represents the [identity of the sender](https://developer.nexmo.com/messaging/sms/guides/custom-sender-id) of the verification request. Depending on the destination of the phone number you are sending the verification SMS to, restrictions might apply.
- * @param "CodeLength" (optional.Int32) -  The length of the verification code.
- * @param "Lg" (optional.String) -  By default, the SMS or text-to-speech (TTS) message is generated in the locale that matches the `number`. For example, the text message or TTS message for a `33*` number is sent in French. Use this parameter to explicitly control the language used for the Verify request. A list of languages is available: <https://developer.nexmo.com/verify/guides/verify-languages>
- * @param "PinExpiry" (optional.Int32) -  How long the generated verification code is valid for, in seconds. When you specify both `pin_expiry` and `next_event_wait` then `pin_expiry` must be an integer multiple of `next_event_wait` otherwise `pin_expiry` is defaulted to equal next_event_wait. See [changing the event timings](https://developer.nexmo.com/verify/guides/changing-default-timings).
- * @param "NextEventWait" (optional.Int32) -  Specifies the wait time in seconds between attempts to deliver the verification code.
- * @param "WorkflowId" (optional.Int32) -  Selects the predefined sequence of SMS and TTS (Text To Speech) actions to use in order to convey the PIN to your user. For example, an id of 1 identifies the workflow SMS - TTS - TTS. For a list of all workflows and their associated ids, please visit the [developer portal](https://developer.nexmo.com/verify/guides/workflows-and-events).
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The response format.
+  - @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
+  - @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
+  - @param number The mobile or landline phone number to verify. Unless you are setting `country` explicitly, this number must be in [E.164](https://en.wikipedia.org/wiki/E.164) format.
+  - @param brand An 18-character alphanumeric string you can use to personalize the verification request SMS body, to help users identify your company or application name. For example: \\\"Your `Acme Inc` PIN is ...\\\"
+  - @param optional nil or *VerifyRequestOpts - Optional Parameters:
+  - @param "Country" (optional.String) -  If you do not provide `number` in international format or you are not sure if `number` is correctly formatted, specify the two-character country code in `country`. Verify will then format the number for you.
+  - @param "SenderId" (optional.String) -  An 11-character alphanumeric string that represents the [identity of the sender](https://developer.nexmo.com/messaging/sms/guides/custom-sender-id) of the verification request. Depending on the destination of the phone number you are sending the verification SMS to, restrictions might apply.
+  - @param "CodeLength" (optional.Int32) -  The length of the verification code.
+  - @param "Lg" (optional.String) -  By default, the SMS or text-to-speech (TTS) message is generated in the locale that matches the `number`. For example, the text message or TTS message for a `33*` number is sent in French. Use this parameter to explicitly control the language used for the Verify request. A list of languages is available: <https://developer.nexmo.com/verify/guides/verify-languages>
+  - @param "PinExpiry" (optional.Int32) -  How long the generated verification code is valid for, in seconds. When you specify both `pin_expiry` and `next_event_wait` then `pin_expiry` must be an integer multiple of `next_event_wait` otherwise `pin_expiry` is defaulted to equal next_event_wait. See [changing the event timings](https://developer.nexmo.com/verify/guides/changing-default-timings).
+  - @param "NextEventWait" (optional.Int32) -  Specifies the wait time in seconds between attempts to deliver the verification code.
+  - @param "WorkflowId" (optional.Int32) -  Selects the predefined sequence of SMS and TTS (Text To Speech) actions to use in order to convey the PIN to your user. For example, an id of 1 identifies the workflow SMS - TTS - TTS. For a list of all workflows and their associated ids, please visit the [developer portal](https://developer.nexmo.com/verify/guides/workflows-and-events).
+
 @return RequestResponse
 */
-func (a *DefaultApiService) VerifyRequest(ctx _context.Context, format string, apiKey string, apiSecret string, number string, brand string, localVarOptionals *VerifyRequestOpts) (RequestResponse, *_nethttp.Response, error) {
+func (a *DefaultApiService) VerifyRequest(ctx context.Context, format string, apiKey string, apiSecret string, number string, brand string, localVarOptionals *VerifyRequestOpts) (RequestResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -272,11 +274,11 @@ func (a *DefaultApiService) VerifyRequest(ctx _context.Context, format string, a
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 	if strlen(brand) > 18 {
 		return localVarReturnValue, nil, reportError("brand must have less than 18 elements")
 	}
@@ -333,7 +335,7 @@ func (a *DefaultApiService) VerifyRequest(ctx _context.Context, format string, a
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -372,25 +374,26 @@ type VerifyRequestWithPSD2Opts struct {
 /*
 VerifyRequestWithPSD2 PSD2 (Payment Services Directive 2) Request
 Use Verify request to generate and send a PIN to your user to authorize a payment: 1. Create a request to send a verification code to your user. 2. Check the &#x60;status&#x60; field in the response to ensure that your request was successful (zero is success). 3. Use the &#x60;request_id&#x60; field in the response for the Verify check. (Please note that XML format is not supported for the Payment Services Directive endpoint at this time.)
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The response format.
- * @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
- * @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
- * @param number The mobile or landline phone number to verify. Unless you are setting `country` explicitly, this number must be in [E.164](https://en.wikipedia.org/wiki/E.164) format.
- * @param payee An alphanumeric string to indicate to the user the name of the recipient that they are confirming a payment to.
- * @param amount The decimal amount of the payment to be confirmed, in Euros
- * @param optional nil or *VerifyRequestWithPSD2Opts - Optional Parameters:
- * @param "Country" (optional.String) -  If you do not provide `number` in international format or you are not sure if `number` is correctly formatted, specify the two-character country code in `country`. Verify will then format the number for you.
- * @param "CodeLength" (optional.Int32) -  The length of the verification code.
- * @param "Lg" (optional.String) -  By default, the SMS or text-to-speech (TTS) message is generated in the locale that matches the `number`. For example, the text message or TTS message for a `33*` number is sent in French. Use this parameter to explicitly control the language used. *Note: Voice calls in English for `bg-bg`, `ee-et`, `ga-ie`, `lv-lv`, `lt-lt`, `mt-mt`, `sk-sk`, `sk-si`
- * @param "PinExpiry" (optional.Int32) -  How long the generated verification code is valid for, in seconds. When you specify both `pin_expiry` and `next_event_wait` then `pin_expiry` must be an integer multiple of `next_event_wait` otherwise `pin_expiry` is defaulted to equal next_event_wait. See [changing the event timings](https://developer.nexmo.com/verify/guides/changing-default-timings).
- * @param "NextEventWait" (optional.Int32) -  Specifies the wait time in seconds between attempts to deliver the verification code.
- * @param "WorkflowId" (optional.Int32) -  Selects the predefined sequence of SMS and TTS (Text To Speech) actions to use in order to convey the PIN to your user. For example, an id of 1 identifies the workflow SMS - TTS - TTS. For a list of all workflows and their associated ids, please visit the [developer portal](https://developer.nexmo.com/verify/guides/workflows-and-events).
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The response format.
+  - @param apiKey You can find your API key in your [account dashboard](https://dashboard.nexmo.com)
+  - @param apiSecret You can find your API secret in your [account dashboard](https://dashboard.nexmo.com)
+  - @param number The mobile or landline phone number to verify. Unless you are setting `country` explicitly, this number must be in [E.164](https://en.wikipedia.org/wiki/E.164) format.
+  - @param payee An alphanumeric string to indicate to the user the name of the recipient that they are confirming a payment to.
+  - @param amount The decimal amount of the payment to be confirmed, in Euros
+  - @param optional nil or *VerifyRequestWithPSD2Opts - Optional Parameters:
+  - @param "Country" (optional.String) -  If you do not provide `number` in international format or you are not sure if `number` is correctly formatted, specify the two-character country code in `country`. Verify will then format the number for you.
+  - @param "CodeLength" (optional.Int32) -  The length of the verification code.
+  - @param "Lg" (optional.String) -  By default, the SMS or text-to-speech (TTS) message is generated in the locale that matches the `number`. For example, the text message or TTS message for a `33*` number is sent in French. Use this parameter to explicitly control the language used. *Note: Voice calls in English for `bg-bg`, `ee-et`, `ga-ie`, `lv-lv`, `lt-lt`, `mt-mt`, `sk-sk`, `sk-si`
+  - @param "PinExpiry" (optional.Int32) -  How long the generated verification code is valid for, in seconds. When you specify both `pin_expiry` and `next_event_wait` then `pin_expiry` must be an integer multiple of `next_event_wait` otherwise `pin_expiry` is defaulted to equal next_event_wait. See [changing the event timings](https://developer.nexmo.com/verify/guides/changing-default-timings).
+  - @param "NextEventWait" (optional.Int32) -  Specifies the wait time in seconds between attempts to deliver the verification code.
+  - @param "WorkflowId" (optional.Int32) -  Selects the predefined sequence of SMS and TTS (Text To Speech) actions to use in order to convey the PIN to your user. For example, an id of 1 identifies the workflow SMS - TTS - TTS. For a list of all workflows and their associated ids, please visit the [developer portal](https://developer.nexmo.com/verify/guides/workflows-and-events).
+
 @return RequestResponse
 */
-func (a *DefaultApiService) VerifyRequestWithPSD2(ctx _context.Context, format string, apiKey string, apiSecret string, number string, payee string, amount float32, localVarOptionals *VerifyRequestWithPSD2Opts) (RequestResponse, *_nethttp.Response, error) {
+func (a *DefaultApiService) VerifyRequestWithPSD2(ctx context.Context, format string, apiKey string, apiSecret string, number string, payee string, amount float32, localVarOptionals *VerifyRequestWithPSD2Opts) (RequestResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPost
+		localVarHTTPMethod   = http.MethodPost
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -400,11 +403,11 @@ func (a *DefaultApiService) VerifyRequestWithPSD2(ctx _context.Context, format s
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/psd2/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 	if strlen(payee) > 18 {
 		return localVarReturnValue, nil, reportError("payee must have less than 18 elements")
 	}
@@ -459,11 +462,11 @@ func (a *DefaultApiService) VerifyRequestWithPSD2(ctx _context.Context, format s
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 
 	// hack to reinstate the body in case we need it
-	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -498,18 +501,19 @@ type VerifySearchOpts struct {
 /*
 VerifySearch Verify Search
 Use Verify search to check the status of past or current verification requests:  1. Send a Verify search request containing the &#x60;request_id&#x60;s of the verification requests you are interested in. 2. Use the &#x60;status&#x60; of each verification request in the &#x60;checks&#x60; array of the response object to determine the outcome.  *Note that this endpoint is available by &#x60;POST&#x60; request as well as &#x60;GET&#x60;.*
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param format The response format.
- * @param apiKey
- * @param apiSecret
- * @param optional nil or *VerifySearchOpts - Optional Parameters:
- * @param "RequestId" (optional.String) -  The `request_id` you received in the Verify Request Response.
- * @param "RequestIds" (optional.Interface of []string) -  More than one `request_id`. Each `request_id` is a new parameter in the Verify Search request.
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param format The response format.
+  - @param apiKey
+  - @param apiSecret
+  - @param optional nil or *VerifySearchOpts - Optional Parameters:
+  - @param "RequestId" (optional.String) -  The `request_id` you received in the Verify Request Response.
+  - @param "RequestIds" (optional.Interface of []string) -  More than one `request_id`. Each `request_id` is a new parameter in the Verify Search request.
+
 @return SearchResponse
 */
-func (a *DefaultApiService) VerifySearch(ctx _context.Context, format string, apiKey string, apiSecret string, localVarOptionals *VerifySearchOpts) (SearchResponse, *_nethttp.Response, error) {
+func (a *DefaultApiService) VerifySearch(ctx context.Context, format string, apiKey string, apiSecret string, localVarOptionals *VerifySearchOpts) (SearchResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodGet
+		localVarHTTPMethod   = http.MethodGet
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -519,11 +523,11 @@ func (a *DefaultApiService) VerifySearch(ctx _context.Context, format string, ap
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/search/{format}"
-	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", _neturl.QueryEscape(parameterToString(format, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"format"+"}", url.QueryEscape(parameterToString(format, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	localVarQueryParams.Add("api_key", parameterToString(apiKey, ""))
 	localVarQueryParams.Add("api_secret", parameterToString(apiSecret, ""))
@@ -568,11 +572,11 @@ func (a *DefaultApiService) VerifySearch(ctx _context.Context, format string, ap
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 
 	// hack to reinstate the body in case we need it
-	localVarHTTPResponse.Body = ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
