@@ -11,17 +11,18 @@
 package voice
 
 import (
-	_context "context"
-	"github.com/antihax/optional"
-	_ioutil "io/ioutil"
-	_nethttp "net/http"
-	_neturl "net/url"
+	"context"
+	"io"
+	"net/http"
+	"net/url"
 	"strings"
+
+	"github.com/antihax/optional"
 )
 
 // Linger please
 var (
-	_ _context.Context
+	_ context.Context
 )
 
 // PlayTTSApiService PlayTTSApi service
@@ -35,15 +36,16 @@ type StartTalkOpts struct {
 /*
 StartTalk Play text to speech into a call
 Play text to speech into a call
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param uuid UUID of the Call Leg
- * @param optional nil or *StartTalkOpts - Optional Parameters:
- * @param "StartTalkRequest" (optional.Interface of StartTalkRequest) -  Action to perform
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param uuid UUID of the Call Leg
+  - @param optional nil or *StartTalkOpts - Optional Parameters:
+  - @param "StartTalkRequest" (optional.Interface of StartTalkRequest) -  Action to perform
+
 @return StartTalkResponse
 */
-func (a *PlayTTSApiService) StartTalk(ctx _context.Context, uuid string, localVarOptionals *StartTalkOpts) (StartTalkResponse, *_nethttp.Response, error) {
+func (a *PlayTTSApiService) StartTalk(ctx context.Context, uuid string, localVarOptionals *StartTalkOpts) (StartTalkResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodPut
+		localVarHTTPMethod   = http.MethodPut
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -53,11 +55,11 @@ func (a *PlayTTSApiService) StartTalk(ctx _context.Context, uuid string, localVa
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/{uuid}/talk"
-	localVarPath = strings.Replace(localVarPath, "{"+"uuid"+"}", _neturl.QueryEscape(parameterToString(uuid, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"uuid"+"}", url.QueryEscape(parameterToString(uuid, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{"application/json"}
@@ -95,7 +97,7 @@ func (a *PlayTTSApiService) StartTalk(ctx _context.Context, uuid string, localVa
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
@@ -124,13 +126,14 @@ func (a *PlayTTSApiService) StartTalk(ctx _context.Context, uuid string, localVa
 /*
 StopTalk Stop text to speech in a call
 Stop text to speech in a call
- * @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- * @param uuid UUID of the Call Leg
+  - @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+  - @param uuid UUID of the Call Leg
+
 @return StopTalkResponse
 */
-func (a *PlayTTSApiService) StopTalk(ctx _context.Context, uuid string) (StopTalkResponse, *_nethttp.Response, error) {
+func (a *PlayTTSApiService) StopTalk(ctx context.Context, uuid string) (StopTalkResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod   = _nethttp.MethodDelete
+		localVarHTTPMethod   = http.MethodDelete
 		localVarPostBody     interface{}
 		localVarFormFileName string
 		localVarFileName     string
@@ -140,11 +143,11 @@ func (a *PlayTTSApiService) StopTalk(ctx _context.Context, uuid string) (StopTal
 
 	// create path and map variables
 	localVarPath := a.client.cfg.BasePath + "/{uuid}/talk"
-	localVarPath = strings.Replace(localVarPath, "{"+"uuid"+"}", _neturl.QueryEscape(parameterToString(uuid, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"uuid"+"}", url.QueryEscape(parameterToString(uuid, "")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{}
@@ -173,7 +176,7 @@ func (a *PlayTTSApiService) StopTalk(ctx _context.Context, uuid string) (StopTal
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
