@@ -4,8 +4,8 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/adevinta/vonage-go-sdk/ncco"
 	"github.com/jarcoal/httpmock"
-	"github.com/vonage/vonage-go-sdk/ncco"
 )
 
 func TestVoiceNewClient(*testing.T) {
